+++ src/eletter/errors.py
@@ -1,11 +1,17 @@
 class Error(Exception):
-    """ The superclass of all custom exceptions raised by ``eletter`` """
+    """
+    .. versionadded:: 0.5.0
+
+    The superclass of all custom exceptions raised by ``eletter``
+    """
 
     pass
 
 
 class DecompositionError(Error, ValueError):
     """
+    .. versionadded:: 0.5.0
+
     Raised when ``eletter`` is asked to decompose an
     `~email.message.EmailMessage` with an unrepresentable
     :mailheader:`Content-Type`
@@ -16,6 +22,8 @@ class DecompositionError(Error, ValueError):
 
 class SimplificationError(Error, ValueError):
     """
+    .. versionadded:: 0.5.0
+
     Raised when ``eletter`` is asked to decompose a message that cannot be
     simplified
     """
@@ -25,6 +33,8 @@ class SimplificationError(Error, ValueError):
 
 class MixedContentError(SimplificationError):
     """
+    .. versionadded:: 0.5.0
+
     Subclass of `SimplificationError` raised when a :mimetype:`multipart/mixed`
     is encountered in which one or more attachments precede a message body
     part; such messages can be forced to be simplified by setting the ``unmix``