+++ docs/changelog.rst
@@ -39,8 +39,8 @@ v0.3.0 (2021-03-11)
 v0.2.0 (2021-03-09)
 -------------------
 - Gave `BytesAttachment` and `FileAttachment` each a `from_file()` classmethod
-- The `from_` and `reply_to` arguments to `compose()` may now be passed lists
-  of addresses
+- The ``from_`` and ``reply_to`` arguments to `compose()` may now be passed
+  lists of addresses
 - Support address groups
 - Added `assemble_content_type()`, `format_addresses()`, and `reply_quote()`
   utility functions